Emit legacy events *after* new events
This is a partial solution. The problem is that event handlers from
packages are getting invoked before event handlers from the editor in
core, which need to be invoked first in order to maintain consistency
with the state of the buffer. If packages attempt to translate buffer
positions before the display buffer is made consistent with the buffer
state following the event they’re handling, exceptions can be raised.

References atom/atom#3192