On transaction undo, restore markers that intersect changes or are moved
In the previous commits, we snapshotted all markers before every
transaction and restored them on undo. This caused markers to not be
in the correct state during buffer change events, which changed our
contract.

Now the state of every marker that is changed destructively by a buffer
change is recorded with the buffer change. Additionally, we also restore
any markers that were explicitly moved during the transaction.

Signed-off-by: Max Brunsfeld <maxbrunsfeld@gmail.com>