Rename “source” to “input” everywhere
Patches are becoming a central concept in this system. A patch is
basically a function that takes input and produces output. This is
clearer terminology than source and target and emphasizes the streaming/
transformational nature of the system.

There’s a chance “input” could be confused with user input, but I don’t
think it’s a big enough chance to not rename stuff.