Supply `getRange` to `scanInRange` iterator
While scanning a large file, we may not want to compute the position
for every result. This is especially important for searches where
there’s an elevated amount of matches and we’re interested only in a
subset of them.

This commit introduces a `getRange` function in the `iterator`
function, which is suited to be a drop-in replacement for `range`. The
logic is the same, but passing a function, or a getter if you will,
enables us to perform all the position math (which is rather expensive)
lazily. :racehorse:

This is still “slow”, however, because we cannot delete the `range`
parameter without an incremental approach, therefore I needed to keep
the eager position computation. As soon as we’ve migrated all the
clients, we can get rid of it and benefit from a significant
performance boost.