Update crash recovery file when buffer modification status changes
When a buffer is modified, we store a crash recovery file. When a buffer
is saved or otherwise returned to an unmodified status, we delete the
recovery file. In a subsequent commit, if a buffer is loaded with a
matching recovery file present, it will report itself as having a crash
recovery available.