WIP: Failing clipping specs
We should really test various clipping and translation scenarios around
empty-line indent guides and end-of-line invisibles.

I wonder if a strategy of translating to a buffer position, seeking the
patch iterator to that buffer position, then translating back could
help out here… That would automatically force us to the beginning of
runs of zero-input-extent regions, but would be a no-op in other cases.

I have a partial solution that follows this strategy, but it would need
to apply inside change regions as well.