Destroy invalid folds after a buffer change
This will happen by marking fold ranges with `{invalidate: ‘inside’}`,
so that when the buffer changes we delete all the invalid folds.
Editing the buffer seems the only way in which folds can be invalidated
(because we don’t expose them), so it should be okay to only destroy
them in response to buffer changes.