Don't reapply flushed changes at the end of a transaction
Previously I thought we could just get away with reapplying all the changes
reported in the `onDidChangeText` event after one or more queries were issued in
the middle of a transaction.

This, however, is simply not possible, because it screws up the expansion of
buffer ranges, and those changes we get in the event won't match the stored
spatial information in the display index.