Perform screen-to-buffer conversion correctly in DisplayMarkerLayer
Previously we were clipping backwards, which can cause problems when the
fold is the latest token in the entire buffer. We cannot use `forward`
as a clipDirection either, because that would make the translation go to
the next line when clipping buffer position in the case of a "normal
fold". Hence, we use the `closest` strategy, which is going to
necessarily clip forward in the current line.