Add DisplayLayer.prototype.softWrapDescriptorForScreenRow
This new private API can be used in `TextEditorPresenter` to
efficiently render line-numbers in the gutter by querying the
`DisplayLayer`'s spatial information just once. We call this a soft-wrap
descriptor, because it describes if and how a row has been soft-wrapped,
including which buffer row such screen row belongs to.