Clarify ownership of MarkerLayers
If a DisplayMarkerLayer is created via DisplayLayer.addMarkerLayer,
which is the normal case, it owns its underlying buffer MarkerLayer and
should destroy it when it is destroyed. However, legacy APIs force us to
allow DisplayMarkerLayers to be constructed for existing buffer
MarkerLayers, in which case we don’t want them to *own* the underlying
MarkerLayer and destroy it when they are destroyed.

Signed-off-by: Antonio Scandurra <as-cii@github.com>