Emit change events on decoration layers and display layers first
After #168, we have observed that sometimes change events are emitted in
a order we don't expect. With this commit we want to ensure that
decoration layers and display layers are always updated in lockstep,
thereby preventing them from having an inconsistent state with respect
to each other.

Another positive aspect of this change is that it should shield against
possible errors happening in third-party 'did-change' event handlers
that get registered before the display layer event handler. In facts,
such errors might cause the change event to not trigger sometimes on the
display layer, and thus to throw errors on future updates to the spatial
lines index.

For these reasons, we are introducing a `registerTextDecorationLayer`
method on `TextBuffer` that needs to be called by decoration layers when
they are initialized. Doing so will enforce that change events are
emitted first on decoration layers, then on display layers and finally
on `TextBuffer.onDidChange` subscribers. This method returns a
`Disposable`, which should be called when the decoration layer is
destroyed in order to stop receiving updates.

An alternative implementation might have computed all the decoration
layers based on the registered display layers on a particular TextBuffer
instance. This, however, would have been somewhat inconvenient because
it would have forced users of `TokenizedBuffer` to always create a
`DisplayLayer`, introducing some unnecessary coupling between the two
objects.