Don't assert about tag order when verifying change events
Decoration layers can insert close tags that change the nesting
order of tags on lines outside of the invalidated range. The
DisplayLayer ensures that tags are properly nested, but does not
guarantee the same tag nesting order across lines when tags
interleave.

This commit makes the randomized tests tolerant of inconsistencies
in tag ordering before and after certain types of decoration changes.

Signed-off-by: Nathan Sobo <nathan@github.com>