Avoid redundant new range computation in setTextInRange
When initially loading large files or processing large insertions of text,
Range.fromText takes significant time to count the lines in a string. We
can avoid this call by waiting to compute the new range until after splicing
into the lines array, at which point we've already counted the lines.