Verify valid state after updating spatial index
We're seeing some rare exceptions in production that seemed to be caused
by the spatial index getting into an invalid state, where it implies the
existence of rows beyond the end of the last buffer row. This carries a
cost but we want to find out if and when this is happening in
production.