Fix preceding boundary calculation for changes inside of folds
If a change is inside a fold and that fold appears at the beginning of a
soft-wrapped line segment, we need to continue iterating upward in our
search for a line boundary. Previously we would stop because the screen
column was 0, but that is invalid if the beginning of the buffer row
isn't on screen due to being folded.

Signed-off-by: Antonio Scandurra <as-cii@github.com>