Clear cached screen lines correctly when range invalidation events occur
Previously, we were mistakenly translating the buffer range to screen
coordinates and invalidating all the cached lines located at that screen
range. This, however, can lead to emitting duplicate lines when folds or
soft wrap occur at the given invalidation range.

As we do elsewhere, with this commit we will find the line boundaries
preceding and following the given buffer range, and splice the screen
lines corresponding to those buffer rows instead.