Parameterize style/className retrieval in `MarkerTextDecorationLayer`
With the previous design it started to feel like we were adding
unnecessary indirection to add a marker text decoration layer. With this
commit we are taking a step back: `addMarkerTextDecorationLayer` got
deleted and the `MarkerTextDecorationLayer` constructor was exposed
instead.

This allows to pass parameters to `MarkerTextDecorationLayer`, which
will stop being in charge of storing and retrieving class names and
inline styles associated to a scope id, leaving that decision up to the
caller.