Fix translation of screen row queries in DisplayMarkerLayer.findMarkers
Previously, when querying a display marker layer with the
`startScreenRow`, `endScreenRow` or `intersectsScreenRowRange`
parameters, we would simply convert the supplied screen rows into buffer
rows and then query the underlying buffer marker layer using the
`startRow`, `endRow` or `intersectsRowRange` parameters.

This behavior was, however, incorrect. In fact, translating a screen row
into a buffer row and using that to perform the query in the buffer
marker layer caused the entire buffer row to be scanned. In presence of
folds or soft-wraps, such query could completely distort the area to
scan that the caller was interested in, and possibly return markers
outside of it or hiding markers inside of it.

With this commit we are changing the translation strategy so that:

* `startScreenRow` queries are translated using the new `startsInRange`
query parameter in the buffer marker layer.
* `endScreenRow` queries are translated using the new `endsInRange`
query parameter in the buffer marker layer.
* `intersectsScreenRowRange` queries are translated using the
`intersectsRange` query parameter in the buffer marker layer.