Don't rely on the history provider to report text change events
Previously, we were relying on `groupChangesSinceCheckpoint` to group
and report a patch representing all the edits that occurred within a
transaction. Depending on the nature of the changes, however, certain
edits may not be included in the history, thus causing the
`did-change-text` and `did-stop-changing` events to not report some
modifications.

With this commit we will stop relying on the history and, instead, every
time the `applyChange` method is called, we will push the change into an
array. Later, when the `emitDidChangeText` method is called (e.g. at the
end of a transaction, or when a file has been reloaded), we will compact
the changes found in such array and emit the event.