Handle {undo: 'skip'} option using .groupLastChanges method
This changes the behavior of undo after calling .setTextInRange with
that option. Previously, the change would not be added to the undo
stack, but this lead to an inconsistent state.

Now, when undoing after calling .setTextInRange with {undo: 'skip'}, the
change is undone along with the preceding change. Hopefully, this
matches the intent of package authors using this option.