Preserve marker identity when restoring a marker after undo/redo
Previously, we would create a *new* marker when restoring a destroyed
marker during an undo or redo. This would leave anyone holding a
reference to the restored marker with a stale copy. This fixes that by
changing our snapshotting mechanism to store a reference to the marker
object when we aren't serializing to disk.