Fix clipping when translating a screen position inside of a hard tab
Previously, there was a logic error in `DisplayLayer.collapseHardTabs`
that was causing that function to not honor the `clipDirection: closest`
parameter when the provided screen position was located inside of a hard
tab.

This commit fixes that logic error by clipping to the next tab stop only
if the given screen position is past the midpoint of the expanded hard
tab.

Signed-off-by: Nathan Sobo <nathan@github.com>